fix(android): 修改getTokenAsync方法的timeout参数类型
- 将getTokenAsync方法中的timeout参数从Int类型改为Double类型
- 在调用HTProtect.getTokenAsync时将timeout转换为Int类型
- 更新新架构接口定义中的参数类型
- 更新旧架构接口定义中的参数类型

+++ android/src/main/java/com/htprotect/HtProtectModule.kt
@@ -106,7 +106,7 @@ class HtProtectModule internal constructor(context: ReactApplicationContext) :
 
 
   @ReactMethod
-  override fun getTokenAsync(businessId: String, timeout: Int, promise: Promise) {
+  override fun getTokenAsync(businessId: String, timeout: Double, promise: Promise) {
     try {
       val myGetTokenCallback = GetTokenCallback { antiCheatResult ->
         if (antiCheatResult.code == AntiCheatResult.OK) {
@@ -120,7 +120,7 @@ class HtProtectModule internal constructor(context: ReactApplicationContext) :
         }
       }
 
-      HTProtect.getTokenAsync(timeout, businessId, myGetTokenCallback)
+      HTProtect.getTokenAsync(timeout.toInt(), businessId, myGetTokenCallback)
     } catch (e: Error) {
       promise.reject("-1", Throwable("[HTProtect]: get token failed."));
     }
